+++ Sources/StreamChat/Repositories/SyncRepository.swift
@@ -5,7 +5,6 @@
 import Foundation
 
 enum SyncError: Error {
-    case localStorageDisabled
     case noNeedToSync
     case tooManyEvents(Error)
     case syncEndpointFailed(Error)
@@ -14,7 +13,7 @@ enum SyncError: Error {
 
     var shouldRetry: Bool {
         switch self {
-        case .localStorageDisabled, .noNeedToSync, .tooManyEvents, .couldNotUpdateUserValue:
+        case .noNeedToSync, .tooManyEvents, .couldNotUpdateUserValue:
             return false
         case .syncEndpointFailed, .failedFetchingChannels:
             return true
@@ -64,6 +63,32 @@ class SyncRepository {
         self.database = database
         self.apiClient = apiClient
     }
+    
+    deinit {
+        cancelRecoveryFlow()
+    }
+    
+    func syncLocalState(completion: @escaping () -> Void) {
+        cancelRecoveryFlow()
+        
+        getUser { [weak self] in
+            guard let currentUser = $0 else {
+                log.error("Current user must exist", subsystems: .offlineSupport)
+                completion()
+                return
+            }
+            
+            guard let lastSyncAt = currentUser.lastSynchedEventDate else {
+                log.info("It's the first session of the current user, skipping recovery flow", subsystems: .offlineSupport)
+                self?.updateUserValue({ $0?.lastSynchedEventDate = Date() }) { _ in
+                    completion()
+                }
+                return
+            }
+            
+            self?.syncLocalState(lastSyncAt: lastSyncAt, completion: completion)
+        }
+    }
 
     /// Syncs the local state with the server to make sure the local database is up to date.
     /// It features queuing, serialization and retries
@@ -76,11 +101,9 @@ class SyncRepository {
     /// 5. Run offline actions requests
     ///
     /// - Parameter completion: A block that will get executed upon completion of the synchronization
-    func syncLocalState(completion: @escaping () -> Void) {
-        operationQueue.cancelAllOperations()
-
+    private func syncLocalState(lastSyncAt: Date, completion: @escaping () -> Void) {
         log.info("Starting to recover offline state", subsystems: .offlineSupport)
-        let context = SyncContext()
+        let context = SyncContext(lastSyncAt: lastSyncAt)
         var operations: [Operation] = []
 
         // Enter recovery mode so no other requests are triggered.
@@ -142,9 +165,26 @@ class SyncRepository {
     /// Syncs the events for the active chat channels using the last sync date.
     /// - Parameter completion: A block that will get executed upon completion of the synchronization
     func syncExistingChannelsEvents(completion: @escaping (Result<[ChannelId], SyncError>) -> Void) {
-        getChannelIds { [weak self] channelIds in
-            self?.syncChannelsEvents(channelIds: channelIds, isRecovery: false, completion: completion)
+        getUser { [weak self] currentUser in
+            guard let lastSyncAt = currentUser?.lastSynchedEventDate else {
+                completion(.failure(.noNeedToSync))
+                return
+            }
+            
+            self?.getChannelIds { channelIds in
+                self?.syncChannelsEvents(
+                    channelIds: channelIds,
+                    lastSyncAt: lastSyncAt,
+                    isRecovery: false,
+                    completion: completion
+                )
+            }
         }
+    }
+    
+    func cancelRecoveryFlow() {
+        operationQueue.cancelAllOperations()
+        apiClient.exitRecoveryMode()
     }
 
     private func getChannelIds(completion: @escaping ([ChannelId]) -> Void) {
@@ -159,38 +199,23 @@ class SyncRepository {
 
     func syncChannelsEvents(
         channelIds: [ChannelId],
+        lastSyncAt: Date,
         isRecovery: Bool,
         completion: @escaping (Result<[ChannelId], SyncError>) -> Void
     ) {
-        guard !channelIds.isEmpty else {
-            completion(.success([]))
+        // In recovery mode, `/sync` should always be called.
+        // Otherwise, the cooldown is checked.
+        guard isRecovery || Date().timeIntervalSince(lastSyncAt) > syncCooldown else {
+            completion(.failure(.noNeedToSync))
             return
         }
 
-        let syncCooldown = self.syncCooldown
-        getUser { [weak self] user in
-            guard let lastSyncAt = user?.lastSynchedEventDate else {
-                // That's the first session of the current user. Bump `lastSyncAt` with current time and return.
-                self?.updateUserValue({
-                    $0?.lastSynchedEventDate = Date()
-                }, completion: { _ in
-                    completion(.failure(.noNeedToSync))
-                })
-                return
-            }
-
-            guard Date().timeIntervalSince(lastSyncAt) > syncCooldown else {
-                completion(.failure(.noNeedToSync))
-                return
-            }
-
-            self?.syncMissingEvents(
-                using: lastSyncAt,
-                channelIds: channelIds,
-                isRecoveryRequest: isRecovery,
-                completion: completion
-            )
-        }
+        syncMissingEvents(
+            using: lastSyncAt,
+            channelIds: channelIds,
+            isRecoveryRequest: isRecovery,
+            completion: completion
+        )
     }
 
     private func getUser(completion: @escaping (CurrentUserDTO?) -> Void) {
@@ -208,6 +233,12 @@ class SyncRepository {
         completion: @escaping (Result<[ChannelId], SyncError>) -> Void
     ) {
         log.info("Synching events for existing channels since \(date)", subsystems: .offlineSupport)
+        
+        guard !channelIds.isEmpty else {
+            completion(.success([]))
+            return
+        }
+        
         let endpoint: Endpoint<MissingEventsPayload> = .missingEvents(since: date, cids: channelIds)
         let requestCompletion: (Result<MissingEventsPayload, Error>) -> Void = { [weak self] result in
             switch result {